crawler/requirements.txt
attrs
cffi
ConcurrentLogHandler
cryptography
cssselect
enum34
funcsigs
future
idna
ipaddress
kafka-python
kazoo
lxml
mock
nose
parsel
pbr
pyasn1
pyasn1-modules
pycparser
PyDispatcher
pyOpenSSL
python-json-logger
PyYAML
queuelib
redis
requests
requests-file
retrying
Scrapy
service-identity
six
testfixtures
tldextract
Twisted
ujson
w3lib
zope.interface
